added dryrun functionality

+++ gpm-to-spotify.sh
@@ -21,27 +21,33 @@ if ! ( which urlencode > /dev/null ); then
     exit 1
 fi
 
-# install gmusicapi
-echo "Installing gmusicapi..."
-pip3 install gmusicapi
+if [[ -f "albums.tmp" && -f "playlists.tmp" ]]; then
+    echo "temp files for playlists & albums already exists, skipping to Spotify upload"
+else
+    # install gmusicapi
+    echo "Installing gmusicapi..."
+    pip3 install gmusicapi
+   
+    # get oauth credentials for GPM
+    read -p "Do you have OAuth credentials already set up for GPM? (y/N) " -r
+    if [[ ! $REPLY =~ ^[Yy]$ ]]; then
+        python3 oauth.py
+    fi
 
-# get oauth credentials for GPM
-read -p "Do you have OAuth credentials already set up for GPM? (y/N) " -r
-if [[ ! $REPLY =~ ^[Yy]$ ]]; then
-    python3 oauth.py
+    # get device ids in a very, very hacky way
+    echo "About to throw an error..."
+    python3 error.py
+    read -p "Please enter one of the device ids listed above: " deviceid
+
+    # get albums from GPM
+    echo "Getting albums and playlists from GPM (this may take some time)..."
+    python3 download.py $deviceid
 fi
 
-# get device ids in a very, very hacky way
-echo "About to throw an error..."
-python3 error.py
-read -p "Please enter one of the device ids listed above: " deviceid
-
-# get albums from GPM
-echo "Getting albums and playlists from GPM (this may take some time)..."
-python3 download.py $deviceid
-
 # Upload music
-./upload.sh
-
-# clean up
-rm albums.tmp playlists.tmp
+if [[ $* == *-d* ]]; then  ## if -d (dryrun) flag was passed
+    ./upload.sh -d
+else
+    ./upload.sh
+    rm albums.tmp playlists.tmp  # clean up
+fi
